split browser tests out to two classes; fix registration test to use an empty users map.
