+++ conf/application-test.conf
@@ -8,3 +8,5 @@ security {
 securesocial{
   ssl=false
 }
+
+databaseToUse = "inmemory" # Options: "inmemory", "ldap"
