+++ test/com/yetu/oauth2provider/base/BaseSpec.scala
@@ -12,12 +12,12 @@ import play.api.test.FakeApplication
  * Class to be extended for all Unit tests
  */
 class BaseSpec extends WordSpec with MustMatchers
-    with GeneratorDrivenPropertyChecks
-    with TestRegistry
-    with BeforeAndAfter
-    with OptionValues
-    with Inside
-    with Inspectors
-    with BaseMethods
-    with DefaultTestVariables
+  with GeneratorDrivenPropertyChecks
+  with TestRegistry
+  with BeforeAndAfter
+  with OptionValues
+  with Inside
+  with Inspectors
+  with BaseMethods
+  with DefaultTestVariables
