+++ test/com/yetu/oauth2provider/browser/BrowserBaseSpec.scala
@@ -1,8 +1,11 @@
 package com.yetu.oauth2provider.browser
 
 import com.yetu.oauth2provider.base.{ TestGlobal, BaseMethods }
+import com.yetu.oauth2provider.oauth2.models.YetuUser
+import com.yetu.oauth2provider.services.data.{MemoryPersonService, MemoryMailTokenService}
 import org.scalatestplus.play.{ HtmlUnitFactory, OneBrowserPerSuite, OneServerPerSuite, PlaySpec }
 import play.api.test.FakeApplication
+import securesocial.core.providers.MailToken
 
 class BrowserBaseSpec extends PlaySpec
     with OneServerPerSuite
@@ -25,5 +28,19 @@ class BrowserBaseSpec extends PlaySpec
       ))
 
   val browserTestUserPassword = "password"
+
+  def clearMailTokensInMemory() = {
+    MemoryMailTokenService.mailTokens = Map[String, MailToken]()
+  }
+
+  def clearUsersFromMemory() = {
+    MemoryPersonService.users = Map[String, YetuUser]()
+  }
+
+  def getMailTokenFromMemory: String = {
+    val tokens = MemoryMailTokenService.mailTokens
+    log(s"TOKENS: $tokens")
+    tokens.head._1
+  }
 
 }
